+++ code/main.go
@@ -23,5 +23,40 @@ func main() {
 		log.Panic(err)
 	}
 	bot := NewBot(os.Getenv("TOKEN"), &db)
+	ticker := time.NewTicker(30 * time.Minute)
+	quit := make(chan struct{})
+	go func() {
+		for {
+			select {
+			case <- ticker.C:
+				calcPlayerIncomeByBiz(&db)
+				log.Printf("ОПА, ВСЕМ БАБКИ!!!")
+			case <- quit:
+				ticker.Stop()
+				return
+			}
+		}
+	}()
+
 	bot.Bot()
+}
+
+func calcPlayerIncomeByBiz(db *DataBase) {
+	players, err := db.GetAllPlayers()
+	if err != nil {
+		log.Fatalf(err.Error())
+	}
+	for _, p := range players {
+		bizes, err := db.GetPlayerBuisnesses(p)
+		if err != nil {
+			log.Fatalf(err.Error())
+		}
+		for _, b := range bizes {
+			t,err := db.GetBusinessTypeById(b.Type)
+			if err != nil {
+				log.Fatalf(err.Error())
+			}
+			db.ChangePlayerMoney(p.PlayerTGID,p.Money + (t.Income * b.Amount))
+		}
+	}
 }