+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## `v0.0.3` Business update (15.08.23)
+
+### Added commands
+
+- /buy 'index' - You can buy business by id
+- /mybizes - List of your businesses
+- /bizes - List of available businesses
+
+### Added features
+
+- Businesses - now you can earn money in AFK, you need just buy some businesss and every 30 minutes you get a income
+
+### Fixed bugs
+
+- GetAllPlayers() function fixed, early it gave an error
+
 ## `v0.0.2` Added some commands, features and fixes (14.08.23)
 
 ### Added commands